feat(wp:plugin): autoloader
