+++ packages/logan-center-plugin/composer.json
@@ -27,6 +27,11 @@
   "require-dev": {
     "squizlabs/php_codesniffer": "3.7.2"
   },
+  "autoload": {
+    "psr-4": {
+      "Klein\\LoganCenter\\": "src/"
+    }
+  },
   "config": {
     "optimize-autoloader": true,
     "preferred-install": "dist",
